+++ Model/CancelShipment.php
@@ -10,6 +10,7 @@ use Exception;
 use Improntus\Uber\Api\WarehouseRepositoryInterface;
 use Improntus\Uber\Helper\Data;
 use Improntus\Uber\Model\Carrier\Uber as UberCarrier;
+use Magento\Framework\Event\ManagerInterface;
 use Magento\Framework\Exception\CouldNotSaveException;
 use Magento\Framework\Exception\InputException;
 use Magento\Framework\Exception\NoSuchEntityException;
@@ -50,8 +51,14 @@ class CancelShipment
     protected OrderShipmentFactory $orderShipmentFactory;
 
     /**
+     * @var ManagerInterface $eventManager
+     */
+    protected ManagerInterface $eventManager;
+
+    /**
      * @param Uber $uber
      * @param Data $helper
+     * @param ManagerInterface $eventManager
      * @param OrderRepository $orderRepository
      * @param array $warehouseRepositories
      * @param OrderShipmentFactory $orderShipmentFactory
@@ -60,6 +67,7 @@ class CancelShipment
     public function __construct(
         Uber                    $uber,
         Data                    $helper,
+        ManagerInterface        $eventManager,
         OrderRepository         $orderRepository,
         array                   $warehouseRepositories,
         OrderShipmentFactory    $orderShipmentFactory,
@@ -67,6 +75,7 @@ class CancelShipment
     ) {
         $this->uber = $uber;
         $this->helper = $helper;
+        $this->eventManager = $eventManager;
         $this->orderRepository = $orderRepository;
         $this->warehouseRepositories = $warehouseRepositories;
         $this->orderShipmentFactory = $orderShipmentFactory;
@@ -136,6 +145,13 @@ class CancelShipment
                     throw new Exception($e->getMessage());
                 }
 
+                // Dispatch Event
+                $this->eventManager->dispatch('uber_shipment_cancel', [
+                    'order' => $order,
+                    'shipment' => $uberOrderShipmentRepository
+                ]);
+
+                // Return Response
                 return $uberResponse;
             }
         }